Move switch case controlling command help message to parser class
